Renaming old Program.cs of each tool
