+++ ToolBox/AlgoSeekFuturesConverter/AlgoSeekFuturesProgram.cs
