+++ ToolBox/AlgoSeekOptionsConverter/AlgoSeekOptionsConverterProgram.cs
