+++ ToolBox/DukascopyDownloader/DukascopyDownloaderProgram.cs
