+++ ToolBox/QuandlBitfinexDownloader/QuandlBitfinexDownloaderProgram.cs
